Refactor auth initialization to centralize in App component
