+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense } from 'react';
+import React, { Suspense, useEffect, useState } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import LandingPage from '@/components/LandingPage';
 import AuthLanding from '@/components/auth/AuthLanding';
@@ -7,11 +7,38 @@ import VerifyOtpPage from '@/components/auth/VerifyOtpPage';
 import AuthGate from '@/components/auth/AuthGate';
 import ToastContainer from '@/components/ui/ToastContainer';
 import { EnvironmentBadge } from '@/components/ui/EnvironmentBadge';
+import { useAuthStore } from '@/stores/authStore';
 
 // LegacyApp はバンドルサイズが大きいため遅延読み込みする
 const LegacyApp = React.lazy(() => import('./LegacyApp'));
 
 const App: React.FC = () => {
+  const [initialized, setInitialized] = useState(false);
+  const { init } = useAuthStore();
+
+  // アプリケーション起動時に一度だけ認証状態を初期化
+  useEffect(() => {
+    const initializeAuth = async () => {
+      console.log('🚀 App: 認証初期化開始');
+      await init();
+      console.log('✅ App: 認証初期化完了');
+      setInitialized(true);
+    };
+    initializeAuth();
+  }, [init]);
+
+  // 初期化中はローディング表示
+  if (!initialized) {
+    return (
+      <div className="w-full h-screen flex items-center justify-center bg-black/70 text-white">
+        <div className="flex flex-col items-center space-y-4">
+          <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-white"></div>
+          <div>Initializing...</div>
+        </div>
+      </div>
+    );
+  }
+
   return (
     <>
       <Suspense
